Split code for readability
